chore: commons auth improvements
