+++ packages/hoppscotch-common/src/components/firebase/Login.vue
@@ -180,7 +180,6 @@ export default defineComponent({
 
       try {
         await platform.auth.signInUserWithGoogle()
-        this.showLoginSuccess()
       } catch (e) {
         console.error(e)
         /*
@@ -203,7 +202,7 @@ export default defineComponent({
       }
 
       if (result.type === "success") {
-        this.showLoginSuccess()
+        // this.showLoginSuccess()
       } else if (result.type === "account-exists-with-different-cred") {
         this.toast.info(`${this.t("auth.account_exists")}`, {
           duration: 0,
@@ -230,7 +229,7 @@ export default defineComponent({
 
       try {
         await platform.auth.signInUserWithMicrosoft()
-        this.showLoginSuccess()
+        // this.showLoginSuccess()
       } catch (e) {
         console.error(e)
         /*
