+++ packages/hoppscotch-common/src/pages/enter.vue
@@ -10,10 +10,7 @@
 import { defineComponent } from "vue"
 import { useI18n } from "@composables/i18n"
 import { initializeFirebase } from "~/helpers/fb"
-
 import { platform } from "~/platform"
-
-import { getLocalConfig, removeLocalConfig } from "~/newstore/localpersistence"
 
 export default defineComponent({
   setup() {
@@ -31,30 +28,14 @@ export default defineComponent({
     initializeFirebase()
   },
   async mounted() {
-    if (platform.auth.isSignInWithEmailLink(window.location.href)) {
-      this.signingInWithEmail = true
+    this.signingInWithEmail = true
 
-      let email = getLocalConfig("emailForSignIn")
-
-      if (!email) {
-        email = window.prompt(
-          "Please provide your email for confirmation"
-        ) as string
-      }
-
-      await platform.auth
-        .signInWithEmailLink(email, window.location.href)
-        .then(() => {
-          removeLocalConfig("emailForSignIn")
-          this.$router.push({ path: "/" })
-        })
-        .catch((e) => {
-          this.signingInWithEmail = false
-          this.error = e.message
-        })
-        .finally(() => {
-          this.signingInWithEmail = false
-        })
+    try {
+      await platform.auth.processMagicLink()
+    } catch (e) {
+      this.error = e.message
+    } finally {
+      this.signingInWithEmail = false
     }
   },
 })
